When starting up a eureka server node, sometimes the server starts up with inbound connections opened but uninitialized outbound connections because of AWS firewall.
While the clients switch over because the server starts up with the EIP binding - the replication of heartbeats to the rest of the eureka server farm fails because of outbound connection failures. This causes the instances that communicate with this eureka server to be expired since the peer eureka nodes do not receive heartbeats and invoke expirations.

This change will ensure the server does not bind to an EIP when it starts up and finds other peer instances until it can send atleast one heartbeat to those.