https://github.com/Netflix/eureka/issues/62
When eureka server startups, it tries to find the unused EIPS based on the list from it neighboring nodes.

If that communication fails for some reason, multiple eureka servers will bind to the same EIP there by snatching it from one another.

AWS does not prevent reassociation in EC2-classic. A better way to handle this edge-case scenario is to find if the an instance id is associated with an elastic IP and if so skip that binding.