Replication optimization where the network traffic between eureka nodes are considerably reduced by executing the replication events as one batch instead of individual requests.
This has both the effect of reducing the number of threads on the eureka server side and also makes the replication lightning fast (where 100s of replication request are executed at the same time as 1 request). This optimization has a great positive effect of scaling eureka servers much much more horizontally making replication and state maintenance a non-issue when the cluster expands to have more nodes.