Full DI for DiscoveryClient
Full DI support for DiscoveryClient with backwards compatibility for
DiscoveryManager.getInstance().  There are default implementations
(cloud) for all DiscoveryClient components and DiscoveryClient can now
be injected directly without the need for a guice module to specify any
of the bindings.
Namespace override is done by binding a String annotated with
@EurekaNamespace in a bootstrap module.