Do not delay on syncUp if there are no peers
syncUp delays up to five times for 30 seconds (by default) if it fails to sync
with its peers.  That's great *if there are peers*, but if there aren't any
(say, you're a developer running a local Eureka), it's simply silly.  Instead
of delaying in that situation, simply immediately return to continue the
startup sequence.

Fixes #42