Simplified resolver.
The current resolver interface has multiple concepts of protocol and ports and provides a stream of changes.
This change, cleans it up by providing a simplistic interface which only provides
host and port and instead of providing a stream of servers, provides a finites list of servers available at that time.

Also cleaned up EurekaClients factory and renamed it to Eureka