Exclude servlet-api jar from eureka.war.
Because of the transitive dependencies we had two servlet-api jar
versions in the dependency tree. Eureka depended directly on
version 2.4 that was properly marked as provided. This version was
however shadowed by the one from the transitive dependency (2.5),
and as it was not marked as provided nor excluded, it was added to
war file. That caused compatibility issue when running on tomcat 8.0.