add @Override toString() in InstanceInfo
Application(Class) can get simple information by @Override the toString().

-> toString() in Application.java

    @Override
    public String toString() {
        return "Application [name=" + name + ", isDirty=" + isDirty
                + ", instances=" + instances + ", shuffledInstances="
                + shuffledInstances + ", instancesMap=" + instancesMap + "]";
    }

It is simple to use some information immediately.

But What is more necessary is the information in InstanceInfo(Class)?
InstanceInfo provides a getter, but it has about 30 member variables.

So it seems necessary to provide the overriden toString() of InstanceInfo