Add stats for DiscoveryClient
We introduce a nested class for `DiscoveryClient` to make it easier for
users to read helpful attributes that can assist in log analysis and
debugging. Most of these attributes are captured by existing metrics:

* `initTimestampMs`: timestamp when the client was initialized
* `localRegistrySize`: number of instances for all applications
* `lastSuccessfulRegistryFetchTimestampMs`: timestamp for last successful fetch
* `lastSuccessfulHeartbeatTimestampMs`: timestamp for last successful heartbeat

We add a new attribute called `initLocalRegistrySize`, which is the
number of instances for all applications read when the client was
initialized. We include a helper method that uses this attribute so
that users can determine whether the client's initial fetch of registry
information succeeded or failed. Note that other accessors are suffixed
with `Ms` (for milliseconds) to denote the time unit.

Additionally, we update the existing `registrySize` with the value from
`initLocalRegistrySize` during initialization instead of waiting until
the first local registry refresh.