Add ability to gate InstanceStatus updates
We add a way for users to define gates when updating their
InstanceStatus that is irrespective of the healthcheck handler
response. For example, if the healthcheck handler returns an
InstanceStatus of `UP` but a gate disallows that status then the
healthcheck status will be overridden and the new status will be
set to `DOWN`.

The interface for a gate has a default method and so can be defined
with a lambda expression.

If there are no gates defined, the default behavior of the discovery
client is unchanged.