Add shouldEnforceFetchRegistryAtInit config
We add a `shouldEnforceFetchRegistryAtInit` configuration to support
users that want the Eureka client to fail fast during initialization.
If `shouldEnforceFetchRegistryAtInit` is set to `true` then the client
will throw an `IllegalStateException` during instance construction if
its initial attempt to fetch of information from the remote Eureka
servers (both primary and backup) is unsuccessful. This setting does
not have an effect if `shouldFetchRegistry` is set to `false`.

We also update `fetchRegistryFromBackup` to return `true` if successful.