Add metrics for instances, tagged with VIP, Status and Expiry
This gives better insight into the current state of the system, and where and changes in load
might be coming from.  I've leveraged the DynamicGauge to capture these metrics, and configured
it to keep data for just 45 seconds. Instances that still exist get refreshed every 30s. With
this setup an instance that does away should hang around for no more than one additional metrics
sampling.