Call size method on Application to avoid copying instance info
DiscoveryClient called getInstancesAsIsFromEureka in a couple locations to get
the number of instances for an application, when Application already has a method
for this purpose that doesn't also incur the overhead of defensively copying.

Additionally, fix a concurrency bug in Application.size, as it was
accessing the underlying instances field without synchronization.